requirements.txt
openai
boto3
python-docx
nest_asyncio
